If !coffee is called with no parameters, helpmebot pours a cup for whomever asked for it.
Simon, I think this should be right (should work).  If it does, great pull it in and I'll update the others that should do this as well; otherwise, please fix any errors and add a little note as to why it wouldn't work like this so I can learn and I'll try again with another (probably "tea" next since this is based on that).  Thanks!